simplify names and types

+++ src/MarsRover/MathExtentions.cs
@@ -3,11 +3,11 @@ namespace MarsRover;
 
 public static class MathExtentions
 {
-    public static byte Wrap(this byte value, byte min, byte max)
+    public static int Wrap(this int value, int min, int max)
     {
         int range = max - min + 1;
         if (range <= 0) throw new ArgumentException("max must be greater than min.");
 
-        return (byte)(((value - min) % range + range) % range + min);
+        return (int)(((value - min) % range + range) % range + min);
     }
 }
